Revert "change fix id type from int to long"
This reverts commit 390ccbbcf9b0b5fb5c9672eb3cf3bba28c72e0a7.

+++ Badcamp/Badcamp.Application/UseCases/EventCases/DeleteEventRequest.cs
@@ -9,6 +9,6 @@ namespace Badcamp.Application.UseCases.EventCases
 {
     public class DeleteEventRequest : IRequest<Response>
     {
-        public long EventId { get; set; }
+        public int EventId { get; set; }
     }
 }
